requirements.txt
# Core
gradio>=4.44,<5
langchain>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-huggingface>=0.1,<0.2
langgraph>=0.4,<0.5
faiss-cpu>=1.7,<2
sentence-transformers>=2.2,<4
numpy<2
requests>=2.28
pydantic>=2,<3
tiktoken>=0.5

# Databricks
databricks-sdk>=0.20
